Moved `absolute:` keyword validation to a trait.
